Rename intgration tests for `Resource`
Tests in `respource_spec.rb` are mostly integration tests. Rather than
trying to refactor/rewrite them en masse, this leaves them there, as
tests for `Resource` (and its backward compatibility). We introduce
`rdf_source_spec.rb` as a home for better unit tests, to be added as
behavior of that module takes shape and its interactors are extracted.